[level 0] Title: 배열의 평균값, Time: 0.02 ms, Memory: 78.2 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.java
@@ -2,7 +2,7 @@ class Solution {
     public double solution(int[] numbers) {
         double answer = 0;
         for (int i = 0; i < numbers.length; i++) {
-            answer += (double) numbers[i];
+            answer += numbers[i];
         }
         return answer / numbers.length;
     }
